Split planning into better submodules
