+++ src/game/lanes_and_cars/planning/current_plan/addable.rs
@@ -3,17 +3,18 @@ use kay::swarm::{Swarm, SubActor, CreateWith};
 use descartes::{Band, P2};
 use ::core::geometry::AnyShape;
 
-use super::{LaneStroke, CurrentPlan};
+use super::CurrentPlan;
+use super::super::lane_stroke::LaneStroke;
 
 #[derive(SubActor, Compact, Clone)]
-pub struct LaneStrokeAddable {
+pub struct Addable {
     _id: Option<ID>,
     stroke: LaneStroke,
 }
 
-impl LaneStrokeAddable {
+impl Addable {
     pub fn new(stroke: LaneStroke) -> Self {
-        LaneStrokeAddable {
+        Addable {
             _id: None,
             stroke: stroke,
         }
@@ -23,7 +24,7 @@ impl LaneStrokeAddable {
 use super::AddToUI;
 use ::core::ui::Add;
 
-impl Recipient<AddToUI> for LaneStrokeAddable {
+impl Recipient<AddToUI> for Addable {
     fn receive(&mut self, msg: &AddToUI) -> Fate {
         match *msg {
             AddToUI => {
@@ -40,7 +41,7 @@ impl Recipient<AddToUI> for LaneStrokeAddable {
 use super::ClearDraggables;
 use ::core::ui::Remove;
 
-impl Recipient<ClearDraggables> for LaneStrokeAddable {
+impl Recipient<ClearDraggables> for Addable {
     fn receive(&mut self, msg: &ClearDraggables) -> Fate {
         match *msg {
             ClearDraggables => {
@@ -54,7 +55,7 @@ impl Recipient<ClearDraggables> for LaneStrokeAddable {
 use ::core::ui::Event3d;
 use super::{AddStroke, Commit};
 
-impl Recipient<Event3d> for LaneStrokeAddable {
+impl Recipient<Event3d> for Addable {
     fn receive(&mut self, msg: &Event3d) -> Fate {
         match *msg {
             Event3d::HoverStarted { .. } |
@@ -73,8 +74,8 @@ impl Recipient<Event3d> for LaneStrokeAddable {
 
 
 pub fn setup() {
-    Swarm::<LaneStrokeAddable>::register_default();
-    Swarm::<LaneStrokeAddable>::handle::<CreateWith<LaneStrokeAddable, AddToUI>>();
-    Swarm::<LaneStrokeAddable>::handle::<ClearDraggables>();
-    Swarm::<LaneStrokeAddable>::handle::<Event3d>();
+    Swarm::<Addable>::register_default();
+    Swarm::<Addable>::handle::<CreateWith<Addable, AddToUI>>();
+    Swarm::<Addable>::handle::<ClearDraggables>();
+    Swarm::<Addable>::handle::<Event3d>();
 }
