+++ src/game/lanes_and_cars/planning/current_plan/rendering.rs
@@ -2,7 +2,8 @@ use kay::{Recipient, Fate, Actor};
 use descartes::{Band, Path, FiniteCurve};
 use monet::{Thing, Instance};
 use ::core::geometry::band_to_thing;
-use super::{CurrentPlan, LaneStroke, DrawingStatus, SelectableStrokeRef};
+use super::{CurrentPlan, DrawingStatus, SelectableStrokeRef};
+use super::super::lane_stroke::LaneStroke;
 
 use monet::SetupInScene;
